Fix parsing of files without extensions
Fixes #65

How embarrassing, file.extname was throwing because in tty file was not
specified, and to add to that, the correct program argument was renamed
to filetype (from ext) and the program was not using the correct
property.

Fixed the issues, and also added unit tests to make sure we cover that
rotten function (getFileType)

Thanks to @friday for reporting it!